# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_NoneType_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[NoneType]",
        "type": "object"
      },
...
      "ApiResponse_RepairResponse_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepairResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[RepairResponse]",
        "type": "object"
      },
...
      "ApiResponse_ServiceResponse_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[ServiceResponse]",
        "type": "object"
      },
...
      "Builder": {
        "enum": [
          "dockerfile",
          "railpack"
        ],
        "title": "Builder",
        "type": "string"
      },
...
      "CreateRepairRequest": {
        "additionalProperties": false,
        "properties": {
          "diagnosisId": {
            "description": "Successful diagnosis belonging to this deployment",
            "exclusiveMinimum": 0.0,
            "title": "Diagnosisid",
            "type": "integer"
          },
          "planIds": {
            "description": "Selected original remediation plan IDs",
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Planids",
            "type": "array"
          }
        },
        "required": [
          "diagnosisId",
          "planIds"
        ],
        "title": "CreateRepairRequest",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "필드별 검증 실패 사유. field 는 camelCase 필드 경로다.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "LatestDeploymentResponse": {
        "description": "서비스 카드에 보여줄 가장 최근 배포 요청. 서비스 상태는 `status` 로 읽는다.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "failureCode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "sourceCommitMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcecommitmessage"
          },
          "sourceSha": {
            "title": "Sourcesha",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DeploymentStatus"
          },
          "triggerType": {
            "$ref": "#/components/schemas/DeploymentTrigger"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "sourceSha",
          "triggerType",
          "createdAt",
          "updatedAt"
        ],
        "title": "LatestDeploymentResponse",
        "type": "object"
      },
...
      "RepairResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "deploymentId": {
            "title": "Deploymentid",
            "type": "integer"
          },
          "diagnosisId": {
            "title": "Diagnosisid",
            "type": "integer"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "planIds": {
            "items": {
              "type": "string"
            },
            "title": "Planids",
            "type": "array"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "sourceSha": {
            "title": "Sourcesha",
            "type": "string"
          },
          "status": {
            "enum": [
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "UNKNOWN_OUTCOME"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "deploymentId",
          "diagnosisId",
          "status",
          "sourceSha",
          "planIds",
          "createdAt"
        ],
        "title": "RepairResponse",
        "type": "object"
      },
...
      "ServiceResponse": {
        "properties": {
          "buildCommand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buildcommand"
          },
          "builder": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Builder"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "dockerfilePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dockerfilepath"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "isAutoDeploy": {
            "title": "Isautodeploy",
            "type": "boolean"
          },
          "latestDeployment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestDeploymentResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "가장 최근 배포 요청. 배포한 적이 없으면 없다."
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "projectId": {
            "title": "Projectid",
            "type": "integer"
          },
          "rootDirectory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rootdirectory"
          },
          "sourceBranch": {
            "title": "Sourcebranch",
            "type": "string"
          },
          "sourceRepositoryUrl": {
            "title": "Sourcerepositoryurl",
            "type": "string"
          },
          "startCommand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startcommand"
          },
          "targetIds": {
            "items": {
              "type": "integer"
            },
            "title": "Targetids",
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "sourceRepositoryUrl",
          "sourceBranch",
          "isAutoDeploy",
          "platform",
          "targetIds",
          "createdAt",
          "updatedAt"
        ],
        "title": "ServiceResponse",
        "type": "object"
      },
...
      "ServiceUpdateRequest": {
        "description": "보낸 필드만 바꾼다. 명시한 null 은 값을 비운다(name·sourceBranch·isAutoDeploy 제외).",
        "properties": {
          "buildCommand": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buildcommand"
          },
          "builder": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Builder"
              },
              {
                "type": "null"
              }
            ]
          },
          "dockerfilePath": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dockerfilepath"
          },
          "isAutoDeploy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isautodeploy"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 63,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "rootDirectory": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rootdirectory"
          },
          "sourceBranch": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcebranch"
          },
          "startCommand": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startcommand"
          },
          "targetIds": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetids"
          }
        },
        "title": "ServiceUpdateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "anydeploy_session",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "description": "CLI 용 세션 토큰",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "모든 JSON 응답은 `ApiResponse` 봉투(`success`·`code`·`message`·`data`)로 감싼다. 필드는 camelCase 다. 인증은 쿠키(웹) 또는 `Authorization: Bearer`(CLI).",
    "title": "AnyDeploy Control API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/services/{service_id}": {
      "delete": {
        "operationId": "delete_service_api_v1_services__service_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND · UPLOAD_NOT_FOUND)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "현재 상태와 충돌한다. 이미 있는 이름이거나 진행 중인 배포가 있다 (CONFLICT · PROJECT_NAME_CONFLICT · SERVICE_NAME_CONFLICT · VARIABLE_CONFLICT · DEPLOYMENT_IN_PROGRESS · DEPLOYMENT_NOT_FAILED · DIAGNOSIS_IN_PROGRESS · DIAGNOSIS_LOGS_UNAVAILABLE · UPLOAD_UNAVAILABLE)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "서비스 삭제 (떠 있는 앱도 함께 내린다)",
        "tags": [
          "services"
        ]
      },
      "get": {
        "operationId": "get_service_api_v1_services__service_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ServiceResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND · UPLOAD_NOT_FOUND)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "서비스 조회",
        "tags": [
          "services"
        ]
      },
      "patch": {
        "operationId": "update_service_api_v1_services__service_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ServiceResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND · UPLOAD_NOT_FOUND)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "현재 상태와 충돌한다. 이미 있는 이름이거나 진행 중인 배포가 있다 (CONFLICT · PROJECT_NAME_CONFLICT · SERVICE_NAME_CONFLICT · VARIABLE_CONFLICT · DEPLOYMENT_IN_PROGRESS · DEPLOYMENT_NOT_FAILED · DIAGNOSIS_IN_PROGRESS · DIAGNOSIS_LOGS_UNAVAILABLE · UPLOAD_UNAVAILABLE)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "외부 시스템(GitHub·Loki·Prometheus·진단 에이전트·S3) 호출에 실패했다 (EXTERNAL_ERROR)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "서비스 설정 수정",
        "tags": [
          "services"
        ]
      }
    },
...
    "/api/v1/services/{service_id}/deployments/{deployment_id}/repairs": {
      "post": {
        "operationId": "create_repair_api_v1_services__service_id__deployments__deployment_id__repairs_post",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RepairResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RepairResponse_"
                }
              }
            },
            "description": "Candidate attempt accepted; poll the repair ID"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "현재 상태와 충돌한다. 이미 있는 이름이거나 진행 중인 배포가 있다 (CONFLICT · PROJECT_NAME_CONFLICT · SERVICE_NAME_CONFLICT · VARIABLE_CONFLICT · DEPLOYMENT_IN_PROGRESS · DEPLOYMENT_NOT_FAILED · DIAGNOSIS_IN_PROGRESS · DIAGNOSIS_LOGS_UNAVAILABLE)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "AI 코드 수정 후보 생성 시작",
        "tags": [
          "repair"
        ]
      }
    },
...
    "/api/v1/services/{service_id}/repairs/{repair_id}": {
      "get": {
        "operationId": "get_repair_api_v1_services__service_id__repairs__repair_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "repair_id",
            "required": true,
            "schema": {
              "title": "Repair Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RepairResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "AI 코드 수정 후보와 상태 조회",
        "tags": [
          "repair"
        ]
      }
    },
    "/api/v1/services/{service_id}/repairs/{repair_id}/artifacts/{name}": {
      "get": {
        "operationId": "get_repair_artifact_api_v1_services__service_id__repairs__repair_id__artifacts__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "repair_id",
            "required": true,
            "schema": {
              "title": "Repair Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "외부 시스템(GitHub·Loki·Prometheus·진단 에이전트) 호출에 실패했다 (EXTERNAL_ERROR)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "무결성 검증한 코드 수정 artifact 다운로드",
        "tags": [
          "repair"
        ]
      }
    },
...
  },
...
}